导出

+++ service/merchant/merchant-api/src/main/java/com/jkm/hss/merchant/entity/MerchantInfoResponse.java
@@ -65,7 +65,7 @@ public class MerchantInfoResponse {
     /**
      * 代理商级别
      */
-    private long firstLevelDealerId;
+    private int firstLevelDealerId;
